% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-criteria.R
\name{ahp.consistency}
\alias{ahp.consistency}
\title{Criteria(or Alternative)-Consistency for AHP}
\usage{
ahp.consistency(CriteriaComparisonMatrix)
}
\arguments{
\item{CriteriaComparisonMatrix}{an nxn comparison matrix. The comparison matrix has integer element ranging 
between 1 and 9. If an element in row i and column j has a value of a then the element in row j and column i 
have the value of 1/a.}
}
\value{
A list of result of AHP consistency test procedure.
}
\description{
This function tests whether a criteria/Alternative comparison matrix is consistent or not. 
The function also returns a priority vector for AHP based decision making.
}
\examples{
K <- matrix(c(
1,7,1/5,1/8,1/2,1/3,1/5,1,
1/7,1,1/8,1/9,1/4,1/5,1/9,1/8,
5,8,1,1/3,4,2,1,1,
8,9,3,1,7,5,3,3,
2,4,1/4,1/7,1,1/2,1/5,1/5,
3,5,1/2,1/5,2,1,1/3,1/3,
5,9,1,1/3,5,3,1,1,
1,8,1,1/3,5,3,1,1
), nrow = 8, byrow = TRUE)
result <- ahp.consistency(K)
print(result$consistent)
#TRUE
}
\references{
Celikbilek Yakup, Cok Kriterli Karar Verme Yontemleri, Aciklamali ve Karsilastirmali
Saglik Bilimleri Uygulamalari ile. Editor: Muhlis Ozdemir, Nobel Kitabevi, Ankara, 2018
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
