% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-criteria.R
\name{ahp.RI}
\alias{ahp.RI}
\title{Random Index for AHP method.}
\usage{
ahp.RI(n)
}
\arguments{
\item{n}{an integer greater or equals to 3.}
}
\value{
Random Index.
}
\description{
This function returns returns the Random Index (RI) for a given n where n >= 3
}
\examples{
ahp.RI(3)
# Returns 0.58
}
\references{
Celikbilek Yakup, Cok Kriterli Karar Verme Yontemleri, Aciklamali ve Karsilastirmali
Saglik Bilimleri Uygulamalari ile. Editor: Muhlis Ozdemir, Nobel Kitabevi, Ankara, 2018
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
