\name{decisionR-package}
\alias{decisionR-package}
\alias{decisionR}
\docType{package}
\title{
\packageTitle{decisionR}
}
\description{
\packageDescription{decisionR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{decisionR}
\packageIndices{decisionR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{decisionR}

Maintainer: \packageMaintainer{decisionR}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
