% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-criteria.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean distance between two vectors}
\usage{
euclidean(p1, p2)
}
\arguments{
\item{p1}{First n-vector}

\item{p2}{Second n-vector}
}
\value{
Euclidean distance
}
\description{
This function simply returns the euclidean distance between two vectors.
}
\examples{
x1 <- c(0,0,0)
x2 <- c(4,3,2)
distance <- euclidean(x1, x2)
print(distance)
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
