% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-criteria.R
\name{euclideanToOrigin}
\alias{euclideanToOrigin}
\title{Euclidean distance between origin and an n-vector}
\usage{
euclideanToOrigin(p1)
}
\arguments{
\item{p1}{An n-vector}
}
\value{
Euclidean distance
}
\description{
This function is a wrapper for \code{euclidean} function in which the first argument
is always origin of euclidean n-space
}
\examples{
x1 <- c(4,3,2)
distance <- euclideanToOrigin(x1)
print(distance)
print(euclidean(x1, c(0,0,0)) == euclideanToOrigin(x1))
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
