% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mabac.R
\name{mabac}
\alias{mabac}
\title{Multi-Attributive Border Approximation area Comparison Method}
\usage{
mabac(mat, w, funcs = NULL)
}
\arguments{
\item{mat}{Decision matrix}

\item{w}{Weights}

\item{funcs}{Vector of functions with elements either min or max}
}
\value{
Returns the results list
}
\description{
Apply Multi-Attributive Border Approximation area Comparison
(mabac) method
for a given decision matrix, weight vector, and vector of functions
with elements either min or max.
}
\examples{
tol <- 0.10
decmat <-
    rbind(
        c(2, 1, 4, 7, 6, 6, 7, 3000),
        c(4, 1, 5, 6, 7, 7, 6, 3500),
        c(3, 2, 6, 6, 5, 6, 8, 4000),
        c(5, 1, 5, 7, 6, 7, 7, 3000),
        c(4, 2, 5, 6, 7, 7, 6, 3000),
        c(3, 2, 6, 6, 6, 6, 6, 3500)
    )
weights <- c(0.293, 0.427, 0.067, 0.027, 0.053, 0.027, 0.053, 0.053)

fns <- c(max, max, max, max, max, max, max, min)

result <- mabac(decmat, weights, fns)
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
