% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-criteria.R
\name{normalize}
\alias{normalize}
\title{Normalizes a vector}
\usage{
normalize(vector)
}
\arguments{
\item{vector}{A vector}
}
\value{
Returns the normalized vector
}
\description{
Normalizes a vector by dividing all of the elements in the vector by the vector norm.
Vector norm is the euclidean distance of the vector with the origin. Squared elements of the vector
 sum up to 1.
}
\examples{
x1 <- c(4,3,2)
x1.normalized <- normalize(x1)
print(x1.normalized)
print(sum(x1.normalized ^ 2) == 1)
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
