% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-criteria.R
\name{prepareDecisionMatrixHeaders}
\alias{prepareDecisionMatrixHeaders}
\title{prepareDecisionMatrixHeaders}
\usage{
prepareDecisionMatrixHeaders(A)
}
\arguments{
\item{A}{Decision matrix}
}
\value{
Same decision matrix with row and column names
}
\description{
This function prepares the matrix headers for a given decision matrix.
The column names are set to Criateri 1, Criteria 2, ..., Criteria M and 
the row names are set to Alternative 1, Alternative 2, ..., Alternative N where 
M is the number of criteria and N is the number of alternatives. This function is not generally
called by the user. The headers are set in the multiple-criteria decision making tools such as Topsis.
}
\examples{
x <- matrix(1:30, nrow = 5, ncol = 6)
decision.matrix <- prepareDecisionMatrixHeaders(x)
print(decision.matrix)
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
