% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saw.R
\name{saw}
\alias{saw}
\title{Simple Additive Weighting Method}
\usage{
saw(mat, w, funcs = NULL)
}
\arguments{
\item{mat}{Decision matrix}

\item{w}{Weights}

\item{funcs}{Vector of functions with elements either min or max}
}
\value{
Returns the results list
}
\description{
Apply Single Additive Weighting (SAW) method
for a given decision matrix, weight vector, and vector of functions
with elements either min or max.
}
\examples{
df <- data.frame(
    c1 = c(25.0, 21, 19, 22),
    c2 = c(65.0, 78, 53, 25),
    c3 = c(7.0, 6, 5, 2),
    c4 = c(20.0, 24, 33, 31)
)
weights <- c(0.25, 0.25, 0.25, 0.25)
fns <- c(max, max, min, max)
result <- saw(df, weights, fns)
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
