% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple-criteria.R
\name{topsis}
\alias{topsis}
\title{Topsis Method}
\usage{
topsis(decisionMatrix, weights)
}
\arguments{
\item{weights}{Weights for criteria (Measure of how important the criteria are.)}

\item{A}{Decision matrix}
}
\value{
A list of results.
}
\description{
This function implements the Topsis method for the multiple-criteria decision problem.
}
\examples{
x <- matrix(1:30, nrow = 5, ncol = 6)
decision.matrix <- prepareDecisionMatrixHeaders(x)
print(decision.matrix)
}
\references{
Celikbilek Yakup, Cok Kriterli Karar Verme Yontemleri, Aciklamali ve Karsilastirmali
Saglik Bilimleri Uygulamalari ile. Editor: Muhlis Ozdemir, Nobel Kitabevi, Ankara, 2018
}
\author{
Mehmet Hakan Satman - mhsatman@istanbul.edu.tr
}
