\name{decisionR-package}
\alias{decisionR-package}
\alias{decisionR}
\docType{package}
\title{
\packageTitle{decisionR}
}
\description{
\packageDescription{decisionR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{decisionR}
\packageIndices{decisionR}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{decisionR}

Maintainer: \packageMaintainer{decisionR}
}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
